+++ stroom-dashboard-client/src/main/java/stroom/dashboard/client/table/TablePresenter.java
@@ -17,11 +17,15 @@
 
 package stroom.dashboard.client.table;
 
-import com.google.gwt.cell.client.Cell;
+import com.google.gwt.cell.client.SafeHtmlCell;
 import com.google.gwt.core.client.GWT;
 import com.google.gwt.dom.client.NativeEvent;
+import com.google.gwt.dom.client.Style;
 import com.google.gwt.dom.client.Style.Unit;
 import com.google.gwt.event.dom.client.ClickEvent;
+import com.google.gwt.safecss.shared.SafeStylesBuilder;
+import com.google.gwt.safehtml.shared.SafeHtml;
+import com.google.gwt.safehtml.shared.SafeHtmlBuilder;
 import com.google.gwt.user.cellview.client.Column;
 import com.google.gwt.view.client.SelectionChangeEvent;
 import com.google.inject.Inject;
@@ -41,10 +45,12 @@ import stroom.dashboard.client.main.ResultComponent;
 import stroom.dashboard.client.main.SearchModel;
 import stroom.dashboard.client.query.QueryPresenter;
 import stroom.dashboard.client.table.TablePresenter.TableView;
+import stroom.dashboard.client.table.cf.ExpressionMatcher;
 import stroom.dashboard.shared.ComponentConfig;
 import stroom.dashboard.shared.ComponentResult;
 import stroom.dashboard.shared.ComponentResultRequest;
 import stroom.dashboard.shared.ComponentSettings;
+import stroom.dashboard.shared.ConditionalFormattingRule;
 import stroom.dashboard.shared.Dashboard;
 import stroom.dashboard.shared.DashboardQueryKey;
 import stroom.dashboard.shared.DataSourceFieldsMap;
@@ -68,16 +74,19 @@ import stroom.dispatch.client.ExportFileCompleteUtil;
 import stroom.document.client.event.DirtyEvent;
 import stroom.document.client.event.DirtyEvent.DirtyHandler;
 import stroom.document.client.event.HasDirtyHandlers;
+import stroom.hyperlink.client.Hyperlink;
 import stroom.node.client.ClientPropertyCache;
 import stroom.node.shared.ClientProperties;
+import stroom.query.api.v2.ExpressionOperator;
+import stroom.query.api.v2.ExpressionTerm;
 import stroom.query.api.v2.ResultRequest.Fetch;
 import stroom.query.shared.v2.ParamUtil;
 import stroom.security.client.ClientSecurityContext;
 import stroom.security.shared.PermissionNames;
 import stroom.svg.client.SvgPresets;
-import stroom.util.client.RandomId;
 import stroom.util.shared.Expander;
 import stroom.util.shared.OffsetRange;
+import stroom.util.shared.RandomId;
 import stroom.widget.button.client.ButtonView;
 import stroom.widget.menu.client.presenter.MenuListPresenter;
 import stroom.widget.popup.client.event.HidePopupEvent;
@@ -103,7 +112,7 @@ public class TablePresenter extends AbstractComponentPresenter<TableView>
 
     private final LocationManager locationManager;
     private final TableResultRequest tableResultRequest = new TableResultRequest(0, 100);
-    private final List<Column<Row, ?>> existingColumns = new ArrayList<>();
+    private final List<Column<TableRow, ?>> existingColumns = new ArrayList<>();
     private final List<HandlerRegistration> searchModelHandlerRegistrations = new ArrayList<>();
     private final ButtonView addFieldButton;
     private final ButtonView downloadButton;
@@ -114,7 +123,7 @@ public class TablePresenter extends AbstractComponentPresenter<TableView>
     private final ClientDispatchAsync dispatcher;
     private final TimeZones timeZones;
     private final FieldsManager fieldsManager;
-    private final DataGridView<Row> dataGrid;
+    private final DataGridView<TableRow> dataGrid;
 
     private int lastExpanderColumnWidth;
     private int currentExpanderColumnWidth;
@@ -124,7 +133,7 @@ public class TablePresenter extends AbstractComponentPresenter<TableView>
     private TableComponentSettings tableSettings;
     private boolean ignoreRangeChange;
     private int[] maxResults = TableComponentSettings.DEFAULT_MAX_RESULTS;
-    private Set<String> usedFieldIds = new HashSet<>();
+    private final Set<String> usedFieldIds = new HashSet<>();
     private List<Field> currentFields = Collections.emptyList();
     private List<String> currentFieldIds = Collections.emptyList();
 
@@ -227,7 +236,7 @@ public class TablePresenter extends AbstractComponentPresenter<TableView>
 
         registerHandler(annotateButton.addClickHandler(event -> {
             if ((event.getNativeButton() & NativeEvent.BUTTON_LEFT) != 0) {
-                annotationManager.showAnnotationMenu(event.getNativeEvent(), currentFields, dataGrid.getSelectionModel().getSelectedItems());
+                annotationManager.showAnnotationMenu(event.getNativeEvent(), getSettings(), dataGrid.getSelectionModel().getSelectedItems());
             }
         }));
     }
@@ -403,7 +412,7 @@ public class TablePresenter extends AbstractComponentPresenter<TableView>
     }
 
     private void enableAnnotate() {
-        final List<EventId> idList = annotationManager.getEventIdList(currentFields, dataGrid.getSelectionModel().getSelectedItems());
+        final List<EventId> idList = annotationManager.getEventIdList(getSettings(), dataGrid.getSelectionModel().getSelectedItems());
         final boolean enabled = idList.size() > 0;
         annotateButton.setEnabled(enabled);
     }
@@ -443,15 +452,14 @@ public class TablePresenter extends AbstractComponentPresenter<TableView>
                 // Don't refresh the table unless the results have changed.
                 final TableResult tableResult = (TableResult) componentResult;
 
-                final List<Row> values = tableResult.getRows();
+                final List<TableRow> values = processData(tableResult.getFields(), tableResult.getRows());
                 final OffsetRange<Integer> valuesRange = tableResult.getResultRange();
 
                 // Only set data in the table if we have got some results and
                 // they have changed.
                 if (valuesRange.getOffset() == 0 || values.size() > 0) {
-//                    updateColumns();
-                    dataGrid.setRowData(valuesRange.getOffset(), values);
                     dataGrid.setRowCount(tableResult.getTotalResults(), true);
+                    dataGrid.setRowData(valuesRange.getOffset(), values);
                 }
 
                 // Enable download of current results.
@@ -460,8 +468,8 @@ public class TablePresenter extends AbstractComponentPresenter<TableView>
                 // Disable download of current results.
                 downloadButton.setEnabled(false);
 
-                dataGrid.setRowData(0, new ArrayList<>());
                 dataGrid.setRowCount(0, true);
+                dataGrid.setRowData(0, new ArrayList<>());
 
                 dataGrid.getSelectionModel().clear();
             }
@@ -473,11 +481,173 @@ public class TablePresenter extends AbstractComponentPresenter<TableView>
         ignoreRangeChange = false;
     }
 
+    private List<TableRow> processData(final List<Field> fields, final List<Row> values) {
+        final List<ConditionalFormattingRule> rules = tableSettings.getConditionalFormattingRules();
+        final Map<String, DataSourceField> fieldMap = tableSettings.getFields()
+                .stream()
+                .collect(Collectors.toMap(Field::getName, field -> new DataSourceField
+                        .Builder()
+                        .name(field.getName())
+                        .type(DataSourceFieldType.INTEGER_FIELD)
+                        .addConditions(
+                                ExpressionTerm.Condition.EQUALS,
+                                ExpressionTerm.Condition.GREATER_THAN,
+                                ExpressionTerm.Condition.GREATER_THAN_OR_EQUAL_TO,
+                                ExpressionTerm.Condition.LESS_THAN,
+                                ExpressionTerm.Condition.LESS_THAN_OR_EQUAL_TO)
+                        .build()));
+        final ExpressionMatcher expressionMatcher = new ExpressionMatcher(fieldMap);
+
+        final List<TableRow> processed = new ArrayList<>(values.size());
+        for (final Row row : values) {
+            boolean hide = false;
+            SafeStylesBuilder rowStyle = new SafeStylesBuilder();
+
+            // Conditional formatting
+            if (rules != null && rules.size() > 0) {
+                try {
+                    ConditionalFormattingRule matchingRule = null;
+                    for (final ConditionalFormattingRule rule : rules) {
+                        try {
+                            if (rule.isEnabled()) {
+                                final Map<String, Object> attributeMap = new HashMap<>();
+                                for (int i = 0; i < fields.size() && i < row.getValues().size(); i++) {
+                                    final Field field = fields.get(i);
+                                    final String value = row.getValues().get(i);
+                                    attributeMap.put(field.getName(), value);
+                                }
+
+                                final ExpressionOperator operator = rule.getExpression();
+                                final boolean match = expressionMatcher.match(attributeMap, operator);
+                                if (match) {
+                                    matchingRule = rule;
+                                    break;
+                                }
+                            }
+                        } catch (final RuntimeException e) {
+                            GWT.log(e.getMessage());
+                        }
+                    }
+                    if (matchingRule != null) {
+                        if (matchingRule.isHide()) {
+                            hide = true;
+                        } else {
+                            if (matchingRule.getBackgroundColor() != null && !matchingRule.getBackgroundColor().isEmpty()) {
+                                rowStyle.trustedBackgroundColor(matchingRule.getBackgroundColor());
+                            }
+                            if (matchingRule.getTextColor() != null && !matchingRule.getTextColor().isEmpty()) {
+                                rowStyle.trustedColor(matchingRule.getTextColor());
+                            }
+                        }
+                    }
+                } catch (final RuntimeException e) {
+                    GWT.log(e.getMessage());
+                }
+            }
+
+            if (!hide) {
+                final Map<String, SafeHtml> safeHtmlMap = new HashMap<>();
+                for (int i = 0; i < fields.size() && i < row.getValues().size(); i++) {
+                    final Field field = fields.get(i);
+                    String value = row.getValues().get(i);
+                    if (value == null) {
+                        value = "";
+                    }
+
+                    final SafeHtmlBuilder builder = new SafeHtmlBuilder();
+                    SafeStylesBuilder stylesBuilder = new SafeStylesBuilder();
+                    stylesBuilder.append(rowStyle.toSafeStyles());
+
+                    // Wrap
+                    if (field.getFormat() != null && field.getFormat().getWrap() != null && field.getFormat().getWrap()) {
+                        stylesBuilder.whiteSpace(Style.WhiteSpace.NORMAL);
+                    }
+                    // Grouped
+                    if (field.getGroup() != null && field.getGroup() >= row.getDepth()) {
+                        stylesBuilder.fontWeight(Style.FontWeight.BOLD);
+                    }
+
+                    String style = stylesBuilder.toSafeStyles().asString();
+                    if (style.length() > 0) {
+                        style = " style=" + style;
+                    }
+                    builder.appendHtmlConstant("<div class=\"cell\"" + style + ">");
+                    append(value, builder);
+                    builder.appendHtmlConstant("</div>");
+
+                    safeHtmlMap.put(field.getId(), builder.toSafeHtml());
+                }
+                processed.add(new TableRow(row.getGroupKey(), row.getDepth(), safeHtmlMap));
+            }
+        }
+        return processed;
+    }
+
+    private void append(final String value, final SafeHtmlBuilder sb) {
+        final List<Object> parts = getParts(value);
+        parts.forEach(p -> {
+            if (p instanceof Hyperlink) {
+                final Hyperlink hyperlink = (Hyperlink) p;
+                if (!hyperlink.getText().trim().isEmpty()) {
+                    sb.appendHtmlConstant("<u link=\"" + hyperlink.toString() + "\">");
+
+//                    if (field != null && field.getFormat() != null && field.getFormat().getType() == Type.DATE_TIME) {
+//                        try {
+//                            long l = Long.parseLong(hyperlink.getText());
+//                            sb.appendEscaped(ClientDateUtil.toISOString(l));
+//                        } catch (final RuntimeException e) {
+//                            sb.appendEscaped(hyperlink.getText());
+//                        }
+//                    } else {
+                    sb.appendEscaped(hyperlink.getText());
+//                    }
+
+                    sb.appendHtmlConstant("</u>");
+                }
+            } else {
+                sb.appendEscaped(p.toString());
+            }
+        });
+    }
+
+    private List<Object> getParts(final String value) {
+        final List<Object> parts = new ArrayList<>();
+
+        final StringBuilder sb = new StringBuilder();
+
+        for (int i = 0; i < value.length(); i++) {
+            final char c = value.charAt(i);
+
+            if (c == '[') {
+                final Hyperlink hyperlink = Hyperlink.create(value, i);
+                if (hyperlink != null) {
+                    if (sb.length() > 0) {
+                        parts.add(sb.toString());
+                        sb.setLength(0);
+                    }
+                    parts.add(hyperlink);
+                    i += hyperlink.toString().length() - 1;
+                } else {
+                    sb.append(c);
+                }
+
+            } else {
+                sb.append(c);
+            }
+        }
+
+        if (sb.length() > 0) {
+            parts.add(sb.toString());
+        }
+
+        return parts;
+    }
+
     private void addExpanderColumn(final int maxDepth) {
         // Expander column.
-        final Column<Row, Expander> column = new Column<Row, Expander>(new ExpanderCell()) {
+        final Column<TableRow, Expander> column = new Column<TableRow, Expander>(new ExpanderCell()) {
             @Override
-            public Expander getValue(final Row row) {
+            public Expander getValue(final TableRow row) {
                 if (row == null) {
                     return null;
                 }
@@ -510,20 +680,10 @@ public class TablePresenter extends AbstractComponentPresenter<TableView>
     }
 
     private void addColumn(final Field field) {
-        final TableCell cell = new TableCell(this, field);
-        final Column<Row, Row> column = new Column<Row, Row>(cell) {
+        final Column<TableRow, SafeHtml> column = new Column<TableRow, SafeHtml>(new SafeHtmlCell()) {
             @Override
-            public Row getValue(final Row row) {
-                return row;
-            }
-
-            @Override
-            public String getCellStyleNames(Cell.Context context, Row object) {
-                if (field.getFormat() != null && field.getFormat().getWrap() != null && field.getFormat().getWrap()) {
-                    return super.getCellStyleNames(context, object) + " " + dataGrid.getResources().dataGridStyle().dataGridCellWrapText();
-                }
-
-                return super.getCellStyleNames(context, object);
+            public SafeHtml getValue(final TableRow object) {
+                return object.getValue(field.getId());
             }
         };
 
@@ -664,7 +824,7 @@ public class TablePresenter extends AbstractComponentPresenter<TableView>
         ensureSpecialFields(IndexConstants.STREAM_ID, IndexConstants.EVENT_ID);
 
         // Remove existing columns.
-        for (final Column<Row, ?> column : existingColumns) {
+        for (final Column<TableRow, ?> column : existingColumns) {
             dataGrid.removeColumn(column);
         }
         existingColumns.clear();
@@ -798,7 +958,7 @@ public class TablePresenter extends AbstractComponentPresenter<TableView>
         return currentFieldIds;
     }
 
-    public List<Row> getSelectedRows() {
+    public List<TableRow> getSelectedRows() {
         return dataGrid.getSelectionModel().getSelectedItems();
     }